requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
clearbit==0.1.7
click==8.1.3
cryptography==39.0.0
dictdiffer==0.9.0
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.0
fastapi==0.89.1
filelock==3.9.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.13
idna==3.4
iso8601==1.1.0
nodeenv==1.7.0
passlib==1.7.4
platformdirs==2.6.2
pre-commit==2.21.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.22.0
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
unicorn==2.0.1.post1
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.17.1
